+++ src/app/modules/core/components/center/dma-center.component.css
@@ -3,4 +3,4 @@
 }
 :host mat-sidenav-container mat-sidenav {
     width: 20em;
-} /*# sourceMappingURL=dma-center.component.css.map */
+} /*# sourceMappingURL=center.component.css.map */
