+++ src/app/modules/core/components/center/dma-center.component.scss
