+++ src/app/modules/core/components/header/dma-header.component.css
@@ -4,4 +4,4 @@
 :host a.navbar-brand img {
     height: 2.4em;
     width: 2.4em;
-} /*# sourceMappingURL=dma-header.component.css.map */
+} /*# sourceMappingURL=header.component.css.map */
