+++ src/app/modules/core/dma-core.module.ts
@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { DmaSharedModule } from '@dma-shared';
-import { DmaCenterComponent } from './components/dma-center/dma-center.component';
-import { DmaHeaderComponent } from './components/dma-header/dma-header.component';
-import { DmaSidebarComponent } from './components/dma-sidebar/dma-sidebar.component';
+import { DmaCenterComponent } from './components/center/dma-center.component';
+import { DmaHeaderComponent } from './components/header/dma-header.component';
+import { DmaSidebarComponent } from './components/sidebar/dma-sidebar.component';
 import { DmaRoutingModule } from './dma-routing.module';
-import { DmaHomeComponent } from './pages/dma-home/dma-home.component';
-import { DmaNotFoundComponent } from './pages/dma-not-found/dma-not-found.component';
+import { DmaHomeComponent } from './pages/home/dma-home.component';
+import { DmaNotFoundComponent } from './pages/not-found/dma-not-found.component';
 
 import { SERIALIZE_INTERCEPTOR_PROVIDER } from '@dma-core/services/interceptors/serialize.interceptor';
