+++ src/app/modules/core/pages/home/dma-home.component.spec.ts
@@ -5,7 +5,7 @@ import { FontAwesomeTestingModule } from '@fortawesome/angular-fontawesome/testi
 import { environment } from '../../../../../environments/environment';
 import { DmaTitleService } from '@dma-shared/services/title-service/dma-title.service';
 import { DmaCollapsingCardComponent } from '@dma-shared/components/collapsing-card/dma-collapsing-card.component';
-import { DmaHomeComponent } from '@dma-core/pages/dma-home/dma-home.component';
+import { DmaHomeComponent } from '@dma-core/pages/home/dma-home.component';
 
 describe('DmaHomeComponent', () => {
     beforeEach(async () => {
